Change multifactor manage/auth/disable routes to use multifactor instead of two-factor
I decided not to rename any methods as the feature is called
two_factor_base, and the methods are really internal.  The route
is the remaining externally visible part using two-factor instead of
multifactor, and since it was added since the last release, it is
possible to change it without affecting backwards compatibility.