Make enabling active_sessions before logout not break global logout
The global logout checkbox was not displayed if enabling
active_sessions before logout previously.  active_sessions
assumed the logout feature was already enabled, so this makes
it a dependency.  It is possible to get it to work without being
a dependency, but it is more work and it is unlikely anyone would
use active_sessions without logout.