Minor tweaks to *_path and *_url accepting query params
Switch from *args to opts={} to avoid array allocation.  It's
unlikely the route_path API will change again now that it accepts
an options hash.

Use unless empty? instead of if any?.  They do the same thing in
this case, but the usage of any? here really means not empty instead
of do are there any entries meeting a criteria.