Ensure webauthn-remove error message is displayed on Bootstrap 4
Bootstrap 4 requires the input field to have class "is-invalid" and to
be a sibling to the "invalid-feedback" element, otherwise the
"invalid-feedback" element won't be displayed.

Since the webauthn remove field error will only be displayed if no
options were selected, we display the error message after the last radio
button. We also use the #input_field_string method to get the
"is-invalid" class back, but modify it to allow for checkboxes as well.