Skip check for error message
The error message is no longer shown if usage is empty (no webauthn
ids for the account).  There is already a check for no webauthn ids
earlier, so you should only hit this case if something removes a
webauthn id between the check at the top of the route and when
displaying the view.  Would could attempt to handle that, but it
doesn't seem worth it.  Just remove the spec line that checks for
it.