Move HTTP basic authentication from #session into #http_basic_auth
Previously, HTTP basic authentication could be performed for any request
that accesses the session (e.g. via #logged_in? or #authenticated?
calls). However, overriding #session has a downside that we cannot use
causing an endless loop.

To remedy this, instead of overriding #session, we move HTTP basic
authentication into a new #http_basic_auth method. This method can be
called in the routing block when HTTP basic authentication is allowed
but not required (which would achieve the previous behaviour):

  route do |r|
    rodauth.http_basic_auth
    r.rodauth
    # ...
  end