Don't call #check_already_logged_in in verify_login_change
When verifying login change, the user is expected to be logged in,
because prior to that they've changed their login, which requires them
to be logged in. This action also does make sense when the user is
logged in.

This allows setting something like

  already_logged_in { redirect login_redirect }

and still having the verify_login_change feature working correctly.