Don't consider account open if in unverified grace period without password
When the account is in the unverified grace period, but its password has
not yet been set (i.e. `verify_account_set_password?` was set to true,
delaying setting password until account verification), a login attempt
to that account won't be allowed.

Currently, the user will get an error message that the password is
invalid. However, this message isn't really accurate, as there is no
password they could enter to be able to login, because the
`#password_match?` method returns false when password doesn't exist. The
default login template does include a button for resending account
verification email, but (a) the user might not notice that link, and (b)
in custom login templates this link might not exist.

A more useful response would be to tell the user they cannot log in
without verifying the account. To do that, we modify the
`#open_account?` method to return true if the account is both in
unverified grace period and has a password. This will evaluate to true
when `#verify_account_set_password?` is left at its default value (which
is false when using verify account grace period), so this only matters
when this setting was manually set to true.