Don't consider account open if in verify account grace period but doesn't have password
When the account is in the unverified grace period, but its password has
not yet been set (i.e. `verify_account_set_password?` was set to true,
delaying setting password until account verification), we should ensure
that we prevent logins for this account.

It seems that login already won't work in this case, because
`#password_match?` method requires that the password hash exists.
However, in this case the user will get an "invalid password" error
message, which is not accurate.

I think it's safer to explicitly handle this scenario and not consider
the account open in this case, which returns the correct error message
that the account is awaiting verification. Essentially, if you're using
verify_account_grace_period and delaying password set, you can only use
the account unverified until the session expires, which is reasonable.