Don't check CSRF token on JSON requests
JSON requests should probably not be expected to provide a CSRF token,
because CSRF protection is normally not part of a JSON API. This makes
JSON requests in Rodauth with both HTML and JWT route versions enabled
(i.e. with json: true) work automatically, the user doesn't need to
disable CSRF protection for JSON routes.