Add "How-To" articles for common scenarios
Currently Rodauth feature documentation is mainly a reference of all
configuration methods, which is good for people already familiar with
Rodauth, but can be a bit overwhelming for people just starting out.

So, we add "How-To" tips for various use cases that are commonly found,
which should hopefully help people achieve what they want more easily.
These guides are inspired by the "How-To" articles in the Devise wiki.