Add password_pepper feature
A password pepper is a secret key that's appended to the password before
it's hashed. This means that, in case an attacker was able to obtain
password hashes from the database, they're not able to use them to crack
the passwords without also knowing the password pepper.

  password_pepper "<long secret key>"

Even though this feature is mainly intended to be used as an alternative
to the approach with database functions (as with that approach it's very
difficult for an attacker to obtain password hashes), we make password
peppers work with it too.

We also implement password pepper rotation functionality, allowing the
user to set a new password pepper in case the current one was
compromised. In this case, password hashes using the old pepper will be
automatically updated to use the new pepper on login.

  password_pepper "new pepper"
  previous_password_peppers ["old pepper", ""]

This also allows introducing a password pepper into an app that already
has password hashes created with no pepper. This can also be viewed as
pepper rotation, where the old pepper is an empty string.

We needed to do extract password matching into separate methods for
disallow_password_reuse feature. One nice result of this is that now all
BCrypt-specific code is behind a method, allowing the user to more
easily switch to a different password hashing algorithm, such as scrypt
or argon2.