Revert "Allow the jwt_refresh route to be used with just a refresh token, without an access token"
This reverts commit 4d648513d87d3e4407e3e9016f2829ca7980c557.

I've decided to revert this because I don't think it is safe.
Without knowing how the current session was logged in (stored
in the access token), we can only do a generic login (it used
the jwt-refresh authentication type), and then it would be
possible to use the same authentication type originally used
to login, and the session would be considered two factor
authenticated when it had only been authenticated by a single
factor.