Support RODA_ROUTE_CSRF=always environment variable in the specs
This checks the csrf token at the top of all route blocks.  By
default, Rodauth's specs do not do this, they only would check
the csrf token for the rodauth actions.  This behavior is more
similar to the RODA_ROUTE_CSRF=no behavior, which uses the
older csrf plugin, which in turn uses the rack_csrf middleware
to perform a similar check for all requests.  The advantage of
this separate check is it will check for request specific
tokens, unlike rack_csrf which doesn't support that.