Add allow_refresh_with_expired_jwt_access_token? configuration method to jwt_refresh feature, for allowing refresh with expired access token
This allows refreshing the jwt with an expired access token, so you
don't need to continually refresh before the access token expires.
Because there should be some limit on the access token, if using
this support, store an hmac based on the refresh token in the
access token, and check that.  This ensures the access token was
based on the same session as the refresh token that created it.

While here, add a basic check when refreshing that the current
session account id matches the account id in the refresh token.
I don't think this presented a security issue, because while
you could get a new refresh token based on the old refresh
token even if the access token was for a different account,
it didn't change the access token's account.  However, it's
still better to check that they match.