Avoid features being loaded multiple times
When loading a tree of dependencies, we need to immediately write down
when we've loaded the dependency, otherwise in this case they can be
loaded multiple times.

It doesn't cause any actual problems, but code users of rodauth.features
and rodauth.routes might rely that every feature is listed exactly once.