Default remember cookie path to "/"
Cookies without a path behave unexpected on requests with multiple path
components. When a "/a/b/c" request sets a cookie, that cookie appears
to be stored for "/a/b" path. This cookie is visible in the inspector on
Firefox, but not on Chrome. Attempting to delete the cookie doesn't
work, because a subsequent request to "/a/b" still sends the cookie (but
not a request to "/" or any other route). See [1] for more details.

This can cause unexpected behaviour when using #load_memory before each
request, as the user that has manually signed out can be autologged back
in from the remember cookie if a request is made to a certain path.

Setting cookie path to "/" resolves this issue. Rails does this by
default for its cookies as well.

[1] https://stackoverflow.com/a/64515458/988414