Fix #clear_session not clearing JWT session with sessions plugin
When sessions plugin is used, `#clear_session` will call
`scope.clear_session`, which unlike `session.clear` will not clear the
`@session` hash.

So in JWT feature we make sure to call `session.clear` in
`#clear_session` in case it wasn't called in the super method.

We also add assertions for setting empty JWT token when `#clear_session`
is called but Rodauth isn't the one returning the response to the client.