Forbid JavaScript from accessing remember cookie
Currently the remember cookie can be retrieved via `document.cookie`,
which opens the surface for attacks if the app using Rodauth has XSS
vulnerability, as it allowing malicious JavaScript code to steal the
remember token and the attacker to login to the user's account.

We mitigate this attack by setting HttpOnly for the remember cookie,
which forbids JavaScript from accessing the cookie. This should
not break code for most people, because the client generally doesn't
need to read this value, it only needs to send the cookie to the server
(which happens automatically).