Add support for use argon2 for password hashing
* Add the ability to migrate from bcrypt to argon2 and vice versa.
* Add ability to change hash cost when used argon2

Also I think there will be no problem, for example, when some users are still using bcrypt and others are already using argon2 after last migration, and we apply new migration with change the cost.

At the moment, I do not know how best to cover changes with tests for cases of migration from argon2 to bcrypt and when changing the cost for argon2, as well as when checking the password through functions in the database.