Refactor the argon2 feature
Raise an error if argon2 1.x is detected, because that uses argon2i
instead of argon2id, which is not compatible with this support.

Change some method and options names to simplify things.

Only require argon2 in the specs that need it, and skip those specs
if argon2 is not installed.

Support RODAUTH_ALWAYS_ARGON2 environment variable to run all specs
with the argon2 feature enabled, to ensure that all code in Rodauth
works with the feature.

Support RODAUTH_NO_ARGON2 environment variable to skip running
argon2 specs.  This is useful with databases where argon2 support was
not added, to test that Rodauth's database functions work correctly
when argon2 support is not enabled.

Rename documentation file to match library file, and update it with a
lot more information, including how to setup the database functions.

Update .ci.gemfile to install argon2, and handle older Ruby versions.
I was not able to get argon2 to work correctly on Ruby 1.9 or 2.0
(those raised "Argon2::ArgonHashFail: Invalid salt size" even though
the size was verified to be correct), so only try to handle back
to Ruby 2.1.  This is mostly academic currently as we are running
only GitHub Actions for CI and have configured that to only run on
Ruby 2.4+, but that may change in the future.

Fixes #147