Support configuring the same feature multiple times
This allows the .configure block to be called multiple times, and where
the same feature can be reconfigured. Previously the following code
wouldn't work:

  plugin :rodauth do
    enable :login
    login_redirect "/dashboard"
  end

  plugin :rodauth do
    enable :login
    use_multi_phase_login? true
  end

This is because when the feature was already enabled, attempting to
enable it again wouldn't do anything, not even extend the configuration
with that feature's configuration. So any feature-specific configuration
wouldn't be available the second time.

This PR makes the feature-loading code run every time, by making the
code that extends list of routes and features idempotent.