Improve Rodauth::Auth inheritance behaviour
When a Rodauth::Auth subclass is inherited, the information about loaded
features and routes is reset. This means it's not possible to use
multiple levels of inheritance reliably, for example:

  class RodauthBase < Rodauth::Auth
    configure do
      # ...
    end
  end

  class RodauthMain < RodauthBase
    # ...
  end

  class RodauthAdmin < RodauthBase
    # ...
  end

Some features such as webauthn_verify_account use the list of all loaded
features to determine behaviour, so this doesn't work if the list is
reset. Likewise, any features that require routing and are enabled in the
superclass but enabled again in the subclass won't have their routes
available in the subclass.

We improve this by initializing attribute values on the Rodauth::Auth
class, and making subclassing making a shallow copy of these values and
assigning them to the subclass.