Cache the configuration object in Rodauth::Auth
Currently it's possible to open a configuration block multiple times for
the same Rodauth::Auth subclass, but the same feature cannot be
configured the second time. This is because a new Configuration object
is used each time, so configuration methods defined after enabling
features don't persist. And because loading the feature is skipped
entirely if it's already been loaded, the new Configuration object won't
be extended with the feature's configuration the second time.

We improve this by caching the Configuration object on the Rodauth::Auth
subclass, and reusing it in each .configure call. Not only does this
allow configuring the same feature the second time, but the feature
doesn't need to be explicitly enabled the second time either.

The ability to configure the same Rodauth::Auth subclass multiple times
can be useful in some scenarios. It also provides the opportunity to
implement proper inheritance behaviour, where an abstract Rodauth::Auth
subclass can configure features to some degree, then any subclasses can
configure the behaviour further.