Add :auth_class plugin option for providing a specific Rodauth::Auth subclass
Currently, Rodauth creates an anonymous Rodauth::Auth subclass when the
plugin is loaded, and uses it for instantiating a Rodauth object.

However, in some cases it's useful for the user to be able to provide
their own Rodauth::Auth subclass, as that gives them more control. This
is currently possible by modifying opts[:rodauths] directly:

  class RodauthMain < Rodauth::Auth
    configure do
      # ...
    end
  end

  class RodauthAdmin < Rodauth::Auth
    configure do
      # ...
    end
  end

  class RodauthApp < Roda
    opts[:rodauths] = {
      nil => RodauthMain,
      :admin => RodauthAdmin,
    }

    plugin :rodauth do
      # just to load the plugin
    end
  end

To make this easier, we add an :auth_class plugin option that can be
used to providing your own Rodauth::Auth subclass.

  class RodauthApp < Roda
    plugin :rodauth, auth_class: RodauthMain do
      # ..
    end

    plugin :rodauth, name: :admin, auth_class: RodauthAdmin do
      # ...
    end
  end