Allow setting additional values to active sessions
This allows overriding the active session insert hash to add additional
values, such as IP addresses and device types.

  active_sessions_insert_hash do
    super.merge(
      ip: request.ip,
      user_agent: request.user_agent,
      ...
    )
  end

It also allows extending the update hash when checking for currently
active session, for keeping the extra information up-to-date:

  active_sessions_update_hash do
    super.merge(
      ip: request.ip,
      user_agent: request.user_agent,
    )
  end

By default, the current session will be updated when
session_inactivity_deadline is set, which is the default. But we also
add a boolean setting for updating the session even if inactivity
deadlines are disabled, to keep the active session data up-to-date
regardless.