Make jwt_refresh feature allow refresh with expired access tokens even if prefix is not set correctly (Fixes #168)
In 2.13.0, we switched from a simple flag on the route to a check
on the route path, to handle cases where the session was needed
before dispatching to the Rodauth route.  Unfortunately, the route
path check doesn't work correctly if the prefix is not set
correctly in the configuration.  Change the check so it allows
decoding an expired access token if either the flag is present
or if the path matches.  This will be faster in the normal case,
and handle cases where the prefix is not set correctly.