Add internal_request feature, for interacting with Rodauth by calling methods
This is expected to be used mostly by adminstrators to automatically
create accounts, change passwords or logins for accounts, and similar
issues.

For example, assuming you've loaded the change_password and
internal_request features, and that your Roda class that
is loading Rodauth is named App, you can change the password
for the account with id 1 using:

  App.rodauth.change_password(account_id: 1, password: 'foobar')

The internal request methods are implemented as class methods
on the Rodauth::Auth subclass (the object returned by App.rodauth).
These methods call methods on a subclass of that class specific
to internal requests.

Internal request methods are implemented by creating a rack
environment hash for the internal request, creating a Roda
instance with it, getting the related Rodauth::Auth instance
for that, and calling a method on that.  In most cases, the
method called is the route handler for the related route, so
the behavior is exactly the same as Rodauth processing a request.

Rodauth will raise a Rodauth::InternalRequestError for errors in
internal requests. In most cases, successful responses will return
nil, but there are internal methods that return non-nil values.
A full description of the internal request methods is included in
the documentation added by this commit.

Implementing this required the addition of a couple minor features:

* Rodauth::Auth classes now have a configuration_name method that
  returns the configuration name associated with the class. They
  also have a configuration method that returns the configuration
  instance for the class.

* Rodauth:::Feature now supports an internal_request_method method
  for specifying which methods are supported as internal request
  methods.

A couple private methods were extracted from specific features in
order to make it easier to implement the related internal request
methods.

Almost all features where an internal request method would make
sense have internal request methods.  The only exceptions are
features related to WebAuthn, which is tied to the browser
fairly closely.  However, if there is a need for internal request
methods related to WebAuthn, those can be added later.