Use configured domain in email links and other URLs
Rodauth allows configuring the domain via the `#domain` configuration
method:

  domain "example.com"

However, this domain wasn't used when generating URLs, because that uses
`Rack::Request#base_url`, which reads the host from the rack env.

This becomes a bigger issue when using the internal_request feature,
which sets the host in rack env to an invalid domain that is expected to
be overridden. Setting the `#domain` gets rid of the error, but e.g.
email links will still use the placeholder domain.

This commit fixes this issue by modifying `#base_url` to use the
configured domain. To retain backwards compatibility with proxies, we
also change `#domain` to read from `Rack::Request#authority`, which is
what `Rack::Request#base_url` read from as well.