Set error status in verification request response
When attempting to login to an unverified account, causing the form for
resending verification email was displayed, currently the response
status was 200 OK. This was probably unintentional, because Rodauth
generally uses 4xx status codes to render error responses. So, we change
this response to use a 4xx status code as well. This makes it work with
Turbo, which expects form submits to either redirect or return a non-2xx
status code.