Assign internal request subclass to a constant
For named auth classes, internal request subclass is still anonymous:

  class RodauthMain < Rodauth::Auth
  end

  class RodauthApp < Roda
    plugin :rodauth, auth_class: RodauthMain do
      enable :internal_request
    end
  end

  RodauthMain.internal_request_eval do
    self.class #=> #<Class:0x0000000143b57680>
  end

We modify the internal_request feature to assign the internal subclass
into a constant if the auth class is not anonymous, so now we get:

  RodauthMain.internal_request_eval do
    self.class #=> RodauthMain::InternalRequest
  end

This improves introspection and error messages. We make the constant
private to prevent it from being accessible.