Rewrite change table and column name guides
Avoid using account_password_hash_column as an example, as it not
recommended to store password hashes in the same table as accounts.

Don't recommend introspecting methods for those ending in table.
That approach is slower and could break if a configuration method
ending with table was added that wasn't related to configuring
table names.