Add maximum_login_bytes, maximum_password_bytes, and maximum_password_length
maximum_login_bytes is being added because for multibyte input, you
probably want to force a maximum number of bytes to whatever can be
stored in the related database column.

Because bcrypt only respects the first 72 bytes of a password, and
password hash calculation speed may be proportional to password
length, it can be useful to support the same minimum and maximum
value for passwords that are used for logins, so add
maximum_password_length and maximum_password_bytes for that.

Add login_too_many_bytes_message, password_too_long_message, and
password_too_many_bytes_message for configuring the error messages.