Make `#require_account` method public
This allows ensuring the account exists in the database before
proceeding with the request. This is useful for preventing errors in
development when the account record has been manually deleted, or if the
account record has been deleted in production for whatever reason, and
the performance impact of fetching the account record on each request is
acceptable.