Add reset_password_notify for emailing users after successful password resets
While users already get notified about requests to reset the
password (since they need to click on the link in the email to
reset the password), previously there was notifying setup for
confirming that a password was changed.  There are cases where
such an email is helpful, and this feature can be used in such
cases.