Work around implicit null byte check added in bcrypt 3.1.18 by checking password requirements before other password checks
An undocumented change in bcrypt 3.1.18 causes it to raise
ArgumentError for passwords with null bytes, due to a change to
use StringValueCStr instead of StringValuePtr in the change to
unlock the GVL while calculating the password hash.

This works around the issue by performing Rodauth's checks on the
password before passing the password to bcrypt.