Ensure routes can be disabled by setting *_route to nil/false
This already worked by accident, because Rodauth would call request.is
with nil/false, which wouldn't match anything, so the matcher would be
skipped.

However, this would result in route hash having a "/" key, which isn't
correct. So, we make sure to skip writing into the route hash when the
*_route method is set to nil/false, and add tests, ensuring that
internal requests still work.

We also add documentation for this feature to the guide for changing
routes.