Do not prefix flash keys with the session key prefix
Flash keys and session keys previously shared code because they
need the same to string conversion if the session doesn't support
symbols.  However, when using a session key prefix, the flash
keys were also prefixed, which is unexpected and undesired.

Separate the session key specific logic from the logic used for
both session keys and flash keys to fix this.

This change is backwards incompatible and can break code using
a session key prefix that expects flash keys to be prefixed.
Applications doing that will need to adjust.