Limit parameter bytesize to 1024 by default, override with max_param_bytesize configuration method
None of the parameters that Rodauth deals with should be over this
size.  All parameters are user submitted and potentially hostile,
so I think it makes sense to enforce a limit.

For more control, users can use the over_max_bytesize_param_value
value configuration method to control parameter values returned
when the parameter value is too long.