Attempt to avoid database errors when invalid tokens are submitted
The vast majority of Rodauth users are using integer values for
database keys, and this change checks that the key values
embedded in tokens are valid integers, avoiding exceptions raised
by the database if they are not.

For Rodauth users not using integer values for database keys,
the convert_token_id configuration method can be used to perform
the appropriate checks or conversions.