Do more checking before trying to get accounts table schema
First, ensure there is a valid Sequel::Database in use.  It is
possible to use Rodauth without Sequel (even examples in the README
for LDAP and facebook).

Second, even if there is a valid Sequel::Database, make sure the
accounts table exists before trying to get the schema for it.
This can fix some applications that only use login and logout,
and override account_from_login and password_match and don't set
accounts_table.