Make `password_hash` method public
There are use cases where one might want to create a password hash
outside of the request. The rodauth-model gem currently calls this
method in the `#password=` setter. The rodauth-rails gem is also now
calling this method in Active Record fixtures, to more easily support
both argon2 and bcrypt along with a potential password pepper.