Add `#account!` method
It's currently possible to call something like this in the routing tree:

  rodauth.account_from_session if rodauth.logged_in?

But that retrieves the account record for all routes, regardless of
whether or not it's needed. And calling it before `r.rodauth` will cause
some Rodauth routes to retrieve the account twice.

So, we add a new `#account!` method that returns the already loaded
account, otherwise it retrieves the account from session if logged in.
We modify a few existing places to use the new method and simplify the
logic.

We skip the login check if the account is already loaded, which allows
calling the method when the Rodauth object has been allocated without a
Roda instance, but has `@account` set (e.g. in background jobs).