Extend remember deadline also while logged in
When extending remember deadline is turned on, this makes the user
remembered for N days after their last activity, as opposed to since
the last time they were autologged in via remember token. This behavior
is probably closer to what the end user expects when asking the app to
remember them.

To avoid executing an update on every request, we wait for a period of
time before extending the remember deadline, 1 hour by default.