Logout no longer active session when remembered
If calling both `rodauth.check_active_session` and `rodauth.load_memory`
in the routing tree, when the session becomes no longer active, the user
will not be logged out if remembered. This is because the user will be
autologged in from the remember cookie right on first request after the
session was cleared. We fix this by also forgetting the user when
clearing the session due to session no longer being active.