Fix db_setup_postgres rake task to work on PostgreSQL 15+
PostgreSQL 15+ removes public create access to the public schema
by default.  Temporarily grant the ph account create access to
the public schema to add the tables it manages, then revoke
the access.

This change is only necessary in PostgreSQL 15+, but should still
work in earlier versions.

Update README for the same reason, to explain to users how to
handle this.