Add WebAuthn Autofill feature
This feature activates autofill UI for passkeys that the user has stored
in the browser on the login page, allowing the user to skip entering
their email address or submitting the login form.

It builds on top of the WebAuthn feature, reusing its login endpoint. It
ships with separate JavaScript that activates the autofill UI on the
login page, and automatically submits the WebAuthn login form when the
user selects their passkey. The WebAuthn login route then retrieves the
user from the WebAuthn credential ID.

This requires requires the autocomplete attribute on login field to include
"webauthn". We now need to generate WebAuthn options without first
identifying the account, so we handle that. The passkeys.dev website
recommends asking for user verification preferred and discoverable
credentials, so we change that as well.