Allow treating user verification in WebAuthn login as 2nd factor
From the FIDO Alliance website[1]:

  Passkeys are kept on a user’s devices (something the user “has”) and —
  if the RP requests User Verification — can only be exercised by the user
  with a biometric or PIN (something the user “is” or ”knows”). Thus,
  authentication with passkeys embodies the core principle of multi-factor
  security.

To allow passwordless login via a WebAuthn to be counted for two factors,
we add a new configuration option that checks whether the user was
verified during the authentication ceremony (e.g. through biometric
identification), and treats that as 2nd factor when using multifactor
authentication.

This is useful for skipping classic 2nd factor methods such as TOTP,
recovery codes or SMS codes, but it's especially useful when the account
has a password or email_auth is enabled, both of which are considered
additional factors and would otherwise be required after WebAuthn login.

[1]: https://fidoalliance.org/passkeys/