Make clearing session work in internal requests
When using Roda sessions plugin, the clear_session method is called on
the Roda object. Similarly, when using rodauth-rails, the reset_session
method is called on the Rails controller instance. During an internal
request, neither of these will actually clear the session, as there the
session hash is stored in an instance variable on the Rodauth object.

We fix this by directly clearing the session hash when in an internal
request. This makes internal_request_eval usage more reliable.