Recommend against using argon2 feature
You would think that because argon2 won the password hashing
competition, it would actually be better than bcrypt.  However,
for typical interactive login situations (generally targetting
under 100ms for password check times), argon2 is significantly
weaker than bcrypt.  Apparently, this was known in cryptographic
circles, but not widely publicized, and I was not aware of it
when I accepted the argon2 feature.

argon2 is significantly stronger than bcrypt for non-interactive
cases, such as when password check times can take multiple seconds,
so it's not that argon2 is never useful, it's just not likely to
be useful in the cases where Rodauth is typically used.