Support hmac_secret rotation via hmac_old_secret configuration method in active_sessions feature
This lays the groundwork for hmac secret rotation, and implements the
rotation support in active_sessions.  When hmac secret rotation is in
use (both hmac_secret and hmac_old_secret in use and not the same),
when looking for existing sessions, active sessions will look for
either hmac.  When updating sessions (done by default, as session
inactivity is handled by default), the current hmac is set in every
UPDATE query.  So the full rotation time for hmac secret for active
sessions need not be greater than the session_inactivity_deadline.